restructuring src/hooks
