+++ src/hooks/auth/dispatchers/index.ts
@@ -1,0 +1,2 @@
+export { default as useDispatchAuthenticate } from './useDispatchAuthenticate';
+export { default as useDispatchSignOut } from './useDispatchSignOut';
