+++ src/hooks/auth/index.ts
@@ -1,0 +1,1 @@
+export { useDispatchAuthenticate, useDispatchSignOut } from './dispatchers';
