+++ src/hooks/blueprint/selectors/useBlueprintAllFields/useBlueprintAllFields.ts
@@ -1,7 +1,7 @@
 import { useSelector, shallowEqual } from 'react-redux';
 import { RootState } from 'src/store/store';
 
-function useBlueprintFields() {
+function useBlueprintAllFields() {
   // TODO: I dont think the shallowEqual does much here but its better than nothing.
   const fields = useSelector(
     ({ blueprint }: RootState) => blueprint.fields,
@@ -10,4 +10,4 @@ function useBlueprintFields() {
   return fields;
 }
 
-export default useBlueprintFields;
+export default useBlueprintAllFields;
