+++ src/hooks/user/dispatchers/index.ts
@@ -1,0 +1,2 @@
+export { default as useDispatchChangePassword } from './useDispatchChangePassword';
+export { default as useDispatchCreateUser } from './useDispatchCreateUser';
